style: SignupFormData을 별도 타입으로 분리하여 타입 순환참조 린트오류 해결
