+++ src/types/form.ts
@@ -1,0 +1,8 @@
+export interface SignupFormData {
+  email: string;
+  password: string;
+  checkPassword: string;
+  contactNumber: string;
+  nickname: string;
+  businessRegistrationNumber: string;
+}
