+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './components';
 export * from './request';
 export * from './response';
 export * from './store';
+export * from './form';
